Create README.md
We are a team of a married couple who have working on  IT projects on the various platforms of the blockchain and  have been working on our entry in to the metaverse together circa in 2018. Our goal is to use the ADAMASffs funds to push ahead emerging women and men colaboration   and create grants for women accross the spectrum in Web 3.0 and to push them there. We will be building an NFT collections supporting developers, scientists and artists. And everyone.
ADAMASffc.  is a decentralized autonomous organization that is focusing on increasing gender diversity in blockchain ecosystems.
ADAMASFFC  is a married couple led Impact  with a focus on empowering and supporting everyone in the metaverse and Web3. Our global community is centered around unique hand drawn NFTs transformed in to digital version on a mission to connect people wherever they are and inspire the next generation of Web 3, equality, fill the gender gap, but also to create a space for new generation partnerships. We want to help all women - the wife and mother, who are mostly underpaid and don't get the recognition they deserve in all fields. We will be donating to artists, developers, singers, actresses, sisters, everyone. Tag a woman in your life that means a lot. But also, to help men. The husband, the father. to help anyone around the world achieve financial, location, and creative independence. Our vision is to help the equal lead, an unconventional life on their own terms.
ADAMASffc mission is to empower women and men, to work as a company -a family company and also a workplace-company, to achieve financial freedom through learning, exploring and building their careers in the Web3 space. Our goal is to grow and diversify the crypto ecosystem by focusing on an end-to-end solution that is fair and frictionless. This includes onboarding, education and direct deployment of women talents into Web3 - as users and leaders.The opportunity. With Web3, we're transitioning into a decentralized economy, which presents a huge opportunity to level the economic playing field. By creating content for a diverse demographic, we want to ensure that women build this new frontier. By simplifying the technical capabilities of Web3 and its potential, we are bridging the gap between the crypto-native and the mainstream. Together. 
                                                                                                           ADAMAS.Feel full not like a fool!